# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Warrior Prayers API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /groups/{groupId}/invites:
    get:
      operationId: listGroupInvites
      tags: [invitations]
      summary: List all invites for a group (admin only)
      security:
        - clerkAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of invites
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GroupInvite"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden — admin role required
    post:
      operationId: createGroupInvite
      tags: [invitations]
      summary: Create a new invite (email or link) — admin only
      security:
        - clerkAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteInput"
      responses:
        "201":
          description: Invite created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupInvite"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
        "403":
          description: Forbidden — admin role required
  /groups/{groupId}/invites/{inviteId}:
    delete:
      operationId: revokeGroupInvite
      tags: [invitations]
      summary: Revoke a pending invite (admin only)
      security:
        - clerkAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
        - name: inviteId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Invite revoked
        "401":
          description: Unauthorized
        "403":
          description: Forbidden — admin role required
        "404":
          description: Invite not found
  /invitations/my:
    get:
      operationId: listMyInvitations
      tags: [invitations]
      summary: List pending invitations for the current user (by email)
      security:
        - clerkAuth: []
      responses:
        "200":
          description: List of pending invitations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MyInvitation"
        "401":
          description: Unauthorized
  /invitations/{token}:
    get:
      operationId: getInviteByToken
      tags: [invitations]
      summary: Get invite details by token (public)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invite details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitePreview"
        "404":
          description: Invite not found or expired
  /invitations/{token}/accept:
    post:
      operationId: acceptInvite
      tags: [invitations]
      summary: Accept an invitation
      security:
        - clerkAuth: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invitation accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupDetail"
        "400":
          description: Invite expired, revoked, or max uses reached
        "401":
          description: Unauthorized
        "404":
          description: Invite not found
  /invitations/{token}/decline:
    post:
      operationId: declineInvite
      tags: [invitations]
      summary: Decline an invitation
      security:
        - clerkAuth: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invitation declined
        "401":
          description: Unauthorized
        "404":
          description: Invite not found
...
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
  schemas:
...
    GroupDetail:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        churchName:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        verse:
          type: ["string", "null"]
        hidePrayerPersonNames:
          type: boolean
        allowCustomCategories:
          type: boolean
        allowComments:
          type: boolean
        allowAnonymousRequests:
          type: boolean
        adminsCanViewAnonymousAuthors:
          type: boolean
        memberCount:
          type: integer
        myRole:
          type: string
          enum: [admin, moderator, member]
        createdByUserId:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - hidePrayerPersonNames
        - allowCustomCategories
        - allowComments
        - allowAnonymousRequests
        - adminsCanViewAnonymousAuthors
        - memberCount
        - myRole
        - createdByUserId
        - createdAt
        - updatedAt
...
    GroupInvite:
      type: object
      properties:
        id:
          type: string
        groupId:
          type: string
        invitedEmail:
          type: ["string", "null"]
        token:
          type: string
        status:
          type: string
          enum: [pending, accepted, declined, expired, revoked]
        expiresAt:
          type: ["string", "null"]
        maxUses:
          type: ["integer", "null"]
        usedCount:
          type: integer
        createdAt:
          type: string
        acceptedAt:
          type: ["string", "null"]
        revokedAt:
          type: ["string", "null"]
        inviteUrl:
          type: string
      required:
        - id
        - groupId
        - token
        - status
        - usedCount
        - createdAt
        - inviteUrl
    InviteInput:
      type: object
      properties:
        invitedEmail:
          type: string
          format: email
        expiresInDays:
          type: integer
          minimum: 1
          maximum: 90
        maxUses:
          type: integer
          minimum: 1
    MyInvitation:
      type: object
      properties:
        id:
          type: string
        token:
          type: string
        groupId:
          type: string
        groupName:
          type: string
        groupDescription:
          type: ["string", "null"]
        invitedByName:
          type: ["string", "null"]
        invitedByEmail:
          type: string
        createdAt:
          type: string
        expiresAt:
          type: ["string", "null"]
      required:
        - id
        - token
        - groupId
        - groupName
        - invitedByEmail
        - createdAt
    InvitePreview:
      type: object
      properties:
        id:
          type: string
        token:
          type: string
        groupId:
          type: string
        groupName:
          type: string
        groupDescription:
          type: ["string", "null"]
        invitedByName:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, accepted, declined, expired, revoked]
        expiresAt:
          type: ["string", "null"]
        isExpired:
          type: boolean
      required:
        - id
        - token
        - groupId
        - groupName
        - status
        - isExpired
...